+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Show more information for runs in error state
 - Improve run stopping in case immediate run stop was not possible
+- Update plugin production dependencies
 
 ## v1.0.1 - 2020-06-05
 ### Fixed