+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## [Unreleased]
+
+## v1.0.2 - 2020-06-17
 ### Fixed
 - Show more information for runs in error state
 - Improve run stopping in case immediate run stop was not possible
 - Update plugin production dependencies
+- Fix internal links in User Guide
 
 ## v1.0.1 - 2020-06-05
 ### Fixed